+++ src/main/java/com/github/dakusui/pcond/core/fluent3/CustomTransformer.java
@@ -1,0 +1,4 @@
+package com.github.dakusui.pcond.core.fluent3;
+
+public interface CustomTransformer {
+}
