+++ src/main/java/com/github/dakusui/pcond/core/fluent3/builtins/DoubleChecker.java
@@ -1,26 +1,30 @@
-package com.github.dakusui.pcond.core.fluent3.typesupports;
+package com.github.dakusui.pcond.core.fluent3.builtins;
 
 
 import com.github.dakusui.pcond.core.fluent3.Matcher;
 
-public interface FloatChecker<
+public interface DoubleChecker<
     R extends Matcher<R, R, OIN, OIN>,
-    OIN> extends
+    OIN
+    > extends
     ComparableNumberChecker<
-        FloatChecker<R, OIN>,
+        DoubleChecker<R, OIN>,
         R,
         OIN,
         Integer> {
 
   class Impl<
       R extends Matcher<R, R, OIN, OIN>,
-      OIN> extends
+      OIN
+      > extends
       Base<
-          FloatChecker<R, OIN>,
+          DoubleChecker<R, OIN>,
           R,
           OIN,
-          Integer>
-      implements FloatChecker<R, OIN> {
+          Integer> implements
+      DoubleChecker<
+          R,
+          OIN> {
     public Impl(OIN rootValue, R root) {
       super(rootValue, root);
     }
