+++ src/main/java/com/github/dakusui/pcond/core/fluent3/builtins/ListChecker.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.pcond.core.fluent3.typesupports;
+package com.github.dakusui.pcond.core.fluent3.builtins;
 
 import com.github.dakusui.pcond.core.fluent3.Checker;
 import com.github.dakusui.pcond.core.fluent3.Matcher;
