+++ src/main/java/com/github/dakusui/pcond/core/fluent3/builtins/LongChecker.java
@@ -1,28 +1,27 @@
-package com.github.dakusui.pcond.core.fluent3.typesupports;
+package com.github.dakusui.pcond.core.fluent3.builtins;
 
 
 import com.github.dakusui.pcond.core.fluent3.Matcher;
 
-public interface DoubleChecker<
+public interface LongChecker<
     R extends Matcher<R, R, OIN, OIN>,
     OIN
     > extends
     ComparableNumberChecker<
-        DoubleChecker<R, OIN>,
+        LongChecker<R, OIN>,
         R,
         OIN,
         Integer> {
 
   class Impl<
       R extends Matcher<R, R, OIN, OIN>,
-      OIN
-      > extends
+      OIN> extends
       Base<
-          DoubleChecker<R, OIN>,
+          LongChecker<R, OIN>,
           R,
           OIN,
           Integer> implements
-      DoubleChecker<
+      LongChecker<
           R,
           OIN> {
     public Impl(OIN rootValue, R root) {
