+++ src/main/java/com/github/dakusui/pcond/core/fluent3/builtins/StreamChecker.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.pcond.core.fluent3.typesupports;
+package com.github.dakusui.pcond.core.fluent3.builtins;
 
 import com.github.dakusui.pcond.core.fluent3.AbstractObjectChecker;
 import com.github.dakusui.pcond.core.fluent3.Matcher;
