--- src/main/java/com/github/dakusui/pcond/core/fluent3/typesupports/CustomTransformer.java
@@ -1,4 +1,0 @@
-package com.github.dakusui.pcond.core.fluent3.typesupports;
-
-public interface CustomTransformer {
-}
